requirements.txt
Django
django-environ
djangorestframework
django-cors-headers
psycopg2
environ